requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
autopep8==1.5
billiard==3.6.2.0
black==19.10b0
boto3==1.11.10
botocore==1.14.10
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.3
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-extensions==2.1.4
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-import-export==2.0.1
django-js-asset==1.2.2
django-rest-passwordreset==0.9.7
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
docutils==0.15.2
drf-schema-adapter==2.0.6
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
importlib-metadata==1.5.0
Inflector==3.0.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.4
kombu==4.6.7
lml==0.0.9
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pathspec==0.8.0
Pillow==5.4.1
psycopg2==2.8.5
psycopg2-binary==2.8.5 
pycodestyle==2.5.0
pycparser==2.19
pydotplus==2.0.2
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
Pygments==2.3.0
PyJWT==1.7.0
pyparsing==2.4.6
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
raven==6.10.0
regex==2020.4.4
requests==2.22.0
s3transfer==0.3.2
sentry-sdk==0.14.1
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
tinycss2==0.6.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
